Mongo native ids are converted to kafka style string ids.
Where an line in the backup is from a record that has a native mongo
id, the value of the '$oid' attribute is extracted and that string value
replaces the native mongo id in the dbObject itself and in the message
wrapper object. For the HBase id, an object { "id": "oid_value" } is used
as this does have to be an object.