Reformat deletes to look like kafka deletes.
In the backups that we import deleted records are the same as regular
records except that the body sits under a '_removed' attribute.

On kafka the records are the same as regular records except that they
have a '@type' of 'MONGO_DELETE'.

When processing deletes in the backup they are now transformed into
something that looks like the record would have done had it come through on
kafka i.e. no top level '_removed' element, but a '@type' of 'MONGO_DELETE'.

We also add a flag into the wrapper to indicate that this transformation
has taken place.